requirements.txt
# Core Slack Bot Dependencies
slack_bolt>=1.18.0
slack_sdk>=3.23.0

# HTTP and API clients
requests>=2.31.0
httpx>=0.24.0

# MCP (Model Context Protocol) dependencies
mcp>=0.5.0
fastmcp>=0.3.0

# AI and Machine Learning
pydantic-ai>=0.0.7
openai>=1.3.0
pydantic>=2.5.0

# Utility libraries
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Streamlit for dashboard
streamlit>=1.28.0

# Optional: Logging and monitoring
structlog>=23.2.0